update to the sidebar
I have added two buttons that will persist through the site to allow the user to access TeacherAide(left) and the sidebar(right) at any point in time. The side bar button is currently linked to the core message box not the AI chatbox. This is because the changes shown were not in the local floder.